drop separate ABC
- fixes #4
